moved the get default endpoint logic to DefaultConfig class
